+++ app/Http/Controllers/WikiEntityImportController.php
@@ -9,9 +9,27 @@ use App\Wiki;
 use App\WikiEntityImport;
 use App\Jobs\WikiEntityImportJob;
 use Carbon\Carbon;
+use Prometheus\CollectorRegistry;
+use Prometheus\Counter;
 
 class WikiEntityImportController extends Controller
 {
+    private Counter $successfulCounter;
+    private Counter $failedCounter;
+
+    public function __construct(CollectorRegistry $registry)
+    {
+        $this->successfulCounter = $registry->getOrRegisterCounter(
+            config('horizon-exporter.namespace'),
+            'wiki_entity_imports_successful',
+            'The number of successful Entity import records.',
+        );
+        $this->failedCounter = $registry->getOrRegisterCounter(
+            config('horizon-exporter.namespace'),
+            'wiki_entity_imports_failed',
+            'The number of failed Entity import records.',
+        );
+    }
     public function get(Request $request): \Illuminate\Http\JsonResponse
     {
         $validatedInput = $request->validate([
@@ -93,6 +111,13 @@ class WikiEntityImportController extends Controller
 
         if ($import->status !== WikiEntityImportStatus::Pending) {
             abort(400, 'Import has to be pending if updated');
+        }
+
+        if ($validatedInput['status'] === WikiEntityImportStatus::Failed->value) {
+            $this->failedCounter->inc();
+        }
+        if ($validatedInput['status'] === WikiEntityImportStatus::Success->value) {
+            $this->successfulCounter->inc();
         }
 
         $import->update([