+++ config/database.php
@@ -157,6 +157,11 @@ return [
             'database' => env('REDIS_CACHE_DB', 1),
         ],
 
+        'metrics' => [
+            'host' => env('REDIS_HOST', '127.0.0.1'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', 6379),
+        ],
     ],
 
 ];